Add type testing to BDV

+++ src/aslm/model/data_sources/bdv_data_source.py
@@ -183,7 +183,9 @@ class BigDataViewerDataSource(DataSource):
                 dtype="float64",
             )
             self.image.create_dataset(
-                f"{setup_group_name}/subdivisions", data=self.subdivisions
+                f"{setup_group_name}/subdivisions",
+                data=self.subdivisions,
+                dtype="int32",
             )
 
         # Create the datasets to populate
